add and fix tests

+++ src/main/java/com/calisapp/repositories/RoutineRepository.java
@@ -13,12 +13,11 @@ import com.calisapp.model.Routine;
 @Configuration
 @Repository
 public interface RoutineRepository extends CrudRepository<Routine, Integer> {
-
 	List<Routine> findAll();
 	
 	@Query(value =  "SELECT * FROM routines r WHERE r.generated_by = :generated_by and r.level = :level", nativeQuery = true)
 	List<Routine> findWithLevel(@Param("level") String level, @Param("generated_by") String generated_by);
 		
 	@Query(value = "SELECT * FROM routines r WHERE r.id in (SELECT routines_id FROM users_routines u WHERE u.user_id = :idUser)", nativeQuery = true)
-	List<Routine> findWithUserId(@Param("idUser") Integer idUser);
+	List<Routine> findWithUserId(@Param("idUser") Long idUser);
 }
